requirements.txt
certifi==2021.5.30
chardet==4.0.0
greenlet==1.1.0
idna==2.10
ldap3==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-ldap==3.3.1
requests==2.25.1
SQLAlchemy==1.4.20
urllib3==1.26.6
